Interests from database
in the proccess of getting data from the database and populating a interests array. The next thing to do is to iterate through this array and use it to dynamically populate a checkbox. there are changes to use http post method here aswell as a regular html form. i found this was not necicary as react allready has an ideal method and variable passing system.