tidying up interests & starting location feature
Just tidying up here, adding some padding around the interests buttons and now creating a better layout for this page. There are further improvements to be made in terms of layout and design but this is like low down on the priority list for me. Stated re-working what i did with the location selection. I am thinking of making it a component that can be called in. That way we can use this in multiple different pages.